refac: 토큰 기반 로그인 추상화
