+++ src/main/java/nextstep/auth/authentication/filter/nonChain/TokenAuthenticationInterceptor.java
@@ -1,0 +1,63 @@
+package nextstep.auth.authentication.filter.nonChain;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import nextstep.auth.application.UserDetailService;
+import nextstep.auth.authentication.AuthenticationException;
+import nextstep.auth.authentication.AuthenticationToken;
+import nextstep.auth.context.Authentication;
+import nextstep.auth.token.JwtTokenProvider;
+import nextstep.auth.token.TokenRequest;
+import nextstep.auth.token.TokenResponse;
+import nextstep.member.domain.LoginMember;
+import org.springframework.http.MediaType;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.stream.Collectors;
+
+public class TokenAuthenticationInterceptor extends NonChainInterceptor {
+    private UserDetailService userDetailService;
+    private JwtTokenProvider jwtTokenProvider;
+
+    public TokenAuthenticationInterceptor(UserDetailService userDetailService, JwtTokenProvider jwtTokenProvider) {
+        this.userDetailService = userDetailService;
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
+
+    @Override
+    protected AuthenticationToken getAuthenticationToken(HttpServletRequest request) throws IOException {
+        String content = request.getReader().lines().collect(Collectors.joining(System.lineSeparator()));
+        TokenRequest tokenRequest = new ObjectMapper().readValue(content, TokenRequest.class);
+
+        String principal = tokenRequest.getEmail();
+        String credentials = tokenRequest.getPassword();
+        return new AuthenticationToken(principal, credentials);
+    }
+
+    @Override
+    protected Authentication getAuthentication(AuthenticationToken authenticationToken) {
+        LoginMember loginMember = userDetailService.loadUserByUsername(authenticationToken.getPrincipal());
+
+        if (loginMember == null) {
+            throw new AuthenticationException();
+        }
+
+        if (!loginMember.checkPassword(authenticationToken.getCredentials())) {
+            throw new AuthenticationException();
+        }
+
+        return new Authentication(loginMember.getEmail(), loginMember.getAuthorities());
+    }
+
+    @Override
+    protected void doAuthentication(Authentication authentication, HttpServletResponse response) throws IOException {
+        String token = jwtTokenProvider.createToken((String) authentication.getPrincipal(), authentication.getAuthorities());
+        TokenResponse tokenResponse = new TokenResponse(token);
+
+        String responseToClient = new ObjectMapper().writeValueAsString(tokenResponse);
+        response.setStatus(HttpServletResponse.SC_OK);
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.getOutputStream().print(responseToClient);
+    }
+}
