--- src/main/java/nextstep/auth/token/TokenAuthenticationInterceptor.java
@@ -1,55 +1,0 @@
-package nextstep.auth.token;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.auth.application.UserDetailService;
-import nextstep.auth.authentication.AuthenticationException;
-import nextstep.member.domain.LoginMember;
-import org.springframework.http.MediaType;
-import org.springframework.web.servlet.HandlerInterceptor;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.util.stream.Collectors;
-
-public class TokenAuthenticationInterceptor implements HandlerInterceptor {
-    private UserDetailService userDetailService;
-    private JwtTokenProvider jwtTokenProvider;
-
-    public TokenAuthenticationInterceptor(UserDetailService userDetailService, JwtTokenProvider jwtTokenProvider) {
-        this.userDetailService = userDetailService;
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
-
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        try {
-
-            String content = request.getReader().lines().collect(Collectors.joining(System.lineSeparator()));
-            TokenRequest tokenRequest = new ObjectMapper().readValue(content, TokenRequest.class);
-
-            String principal = tokenRequest.getEmail();
-            String credentials = tokenRequest.getPassword();
-
-            LoginMember loginMember = userDetailService.loadUserByUsername(principal);
-
-            if (loginMember == null) {
-                throw new AuthenticationException();
-            }
-
-            if (!loginMember.checkPassword(credentials)) {
-                throw new AuthenticationException();
-            }
-
-            String token = jwtTokenProvider.createToken(loginMember.getEmail(), loginMember.getAuthorities());
-            TokenResponse tokenResponse = new TokenResponse(token);
-
-            String responseToClient = new ObjectMapper().writeValueAsString(tokenResponse);
-            response.setStatus(HttpServletResponse.SC_OK);
-            response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-            response.getOutputStream().print(responseToClient);
-        } catch (Exception e) {
-            return false;
-        }
-        return false;
-    }
-}
